Split "AcuantCamera" from "AcuantCaptureCanvas" for managing Acuant lifecycle
**Why**: As currently implemented, the AcuantCaptureCanvas component must remain mounted for the full duration of an Acuant capture, since it's currently responsible for managing callbacks. Since the canvas is a UI component rendered within the FullScreen dialog, this has limited our ability to show different UI states for when an image has been captured, but hasn't yet finished cropping. To be able to improve the responsiveness of the interaction, we can instead split this into two component: One which manages the Acuant lifecycle (AcuantCamera) and one which displays the UI associated with the canvas rendering (AcuantCaptureCanvas).
